Simple parser (#1)
The parser is implemented using Pest and Pest-consume.

We can parse the full Morel grammar, and we can execute all
of Morel Java's .smli tests (just parsing expressions, not
validating or executing).

The shell and parser are still not great at figuring out
when a statement ends. We have fixed #4 but #5 remains; to
work around this, various `.smli` files now have comments
after lines that end in ';', and in other places we have
removed the ';'. Those semicolons are optional but valid,
and we should restore them in the tests when we support
parsing them.

Also, we cannot yet execute negative tests, where the
expression has a syntax error and is supposed to give a
particular error message.