Unifier (#6)
Add a unifier. Given pairs of terms, the unifier finds a
substitution for variables that makes those terms
structurally equivalent.

The unifier will be the basis for the type-resolution
algorithm (see #7). Type-constraints (necessary to support
operator overloading) are not implemented yet.

The unifier is based on Morel Java's
[class Unifier](https://github.com/hydromatic/morel/blob/main/src/main/java/net/hydromatic/morel/util/Unifier.java).
and implements an algorithm due to Martelli, Montanari
(1976) and Paterson, Wegman (1978).

Fixes #6