Non-recursive functions
This commit allows you to declare functions (via a 'let
fun' or 'let val = fn', but without the 'rec' keyword that
would allow a function to reference itself) and to execute
them.

For evaluation, we have a new `struct Frame` that holds the
values of variables in the current stack frame. A stack
frame is created on entry to a function. It is sized based
on the number of variables (atomic patterns) defined in the
code of that function.

The evaluation environment (`struct EvalEnv`) is kept
separate from `Frame` because of lifetime and mutability
issues, even though it unfortunately means passing an extra
parameter to the `eval_xxx` family of functions.