Add Morel Rust language reference
Morel Rust is a subset of Morel Java; accordingly, the Rust
reference is a subset of the Java reference. In the future,
as we implement built-in functions, and features such as
queries, exceptions and overloading, we will paste more
content from the Java to the Rust reference.

Silence Lint complaints about long lines in HTML `<pre>`
sections and Markdown tables.