`String` structure
When encoding record types are as sequence names in the
unifier, switch from colons to backticks. Previously the
record type `{a: int, b: bool}` would be encoded as
"record:a:b", but now it is "record`a`b". Colons caused
problems for the cons operator, `op ::`.

Also add `Char.compare` and `Char.toLower`.