Exceptions
Add the first exception, `Chr`, and a mechanism to propagate
exceptions up the call stack. An exception needs a location in
the code, and therefore we add Span to `core::expr::Apply` and
to certain variants of `Code`.

Also implement functions `Char.chr` and `List.tabulate`:
 * `Char.chr: int -> char`
   Converts an integer to a character.
 * `List.tabulate: int * (int -> 'a) -> 'a list`
   Creates a list by applying a function to indices 0 through
   n-1.

Conveniently, `chr` throws exception `Chr`, which allows us to
test exceptions.