When encoding record types in unifier, quote field names if necessary
Switch encoding from backtick separator to colon separator
with quoting, to be consistent with the Java implementation.

Without this fix, we had trouble encoding field names that
contain a colon -- for example the `::` (cons) field of the
`List` structure -- because colon is used as the separator
in record type encoding. After the fix, record types are
encoded as "record:field1:field2:..." with fields containing
colons quoted as "record:field1:`field:with:colon`:field3".

Add methods `split_quoted` and `join_quoted` in TypeResolver, and use
them in `field_list` and `record_label_from_set`. These methods handle
proper quoting of field names that contain the separator character.

Port of morel-java commit 7b625f2.