Start 'Morel in the browser' (#13)
Fixes #13

To make it easy to call Morel from WASM, we need a simple
interface to the shell. We already have `process_line`,
which takes a string and returns a string. It is stateful,
so variables assigned in one statement are available in the
next statement:

```sml
  let mut shell = Shell::new(&[]);
  let mut result;

  let in_1 = "val x = 5\n\
      and y = 6\n";
  let out_1 = "> val x = 5 : int\n\
      > val y = 6 : int\n";
  result = shell.process_statement(in_1, None).unwrap();
  assert_eq!(result, out_1);

  let in_2 = "x + y\n";
  let out_2 = "> val it = 11 : int\n";
  result = shell.process_statement(in_2, None).unwrap();
  assert_eq!(result, out_2);
```

The current API is string-in, string-out. The input must
be a single declaration or expression (without a
semicolon). The output prints a value and its type. Future
enhancements could be multi-statement input (separated by
semicolons) and output as a structured value.