Optimize `exists` and `forall` queries with short-circuit evaluation (#25)
Fixes #25

Before this change, the `RowSink` evaluation mechanism, and
pipelines that end in `CollectRowSink`, forces `exists`
queries to materialize the entire result set before checking
if it's non-empty. This wastes memory and CPU.

This change adds `ExistsRowSink`, which doesn't keep a list
of rows, just a boolean value. After accepting the first row,
`ExistsRowSink` signals completion by returning a new error
code, `EarlyReturn`. (In Java, using exceptions for control
flow is considered an anti-pattern because there is a
signficant runtime cost, but not so in Rust.) Using an error
causes the signal to bubble up to the producer at the
beginning of the pipeline without writing extra code.

`EarlyReturn` is advisory and idempotent. It indicates that
sending more rows will not affect the query result, but it
is OK for the producer to ignore it and send more rows.

`forall` queries benefit from this mechanism because `forall`
can be rewritten to `not exists`. We do not need to create a
`ForallRowSink`.

`take` queries also benefit from this mechanism.
`TakeRowSink` sends `EarlyReturn` after the limit has been
reached.

Previously we would represent `exists` and `forall` queries
by wrapping them in calls to `Relational.nonEmpty` and
`Relational.empty` functions. But that moved the logic to a
place not seen by the pipeline builder. So we add a new
`StepKind` variant in core, `Exists`; if it appears, it must
be as the last step, and it directs the pipeline builder to
allocate an `ExistsRowSink` rathe than the usual
`CollectRowSink`.