Correct signature of `List.except`, `List.intersect`, `Sys.unset` functions, and remove `Bag.collate`
Most of the functions in the `List` structure make sense in
`Bag` -- even `hd` and `tl` if you think of them as picking
an arbitrary element -- but `collate` is meaningless with
unordered arguments. So we remove it.

Also improve how types are displayed, for the benefit of
`Sys.env`;

Make `Bool.if` an abstract function. We use it only for type
derivation. It is implemented by translaing to `case`, which
is also lazy; a function that eagerly evaluates its
arguments just wouldn't work.